requirements.txt
requests==2.31.0
python-dotenv==1.0.0
w1thermsensor==2.3.0
RPi.GPIO==0.7.1